move copy functions to input_copy
